New data source concept:  "connectors". Allows writing decoupled data sources that can fetch data of general interest, without hard-coding how to align it with a particular thing. This takes advantage of the new `chain.raw` property.
Apply this to burden tests, which we want to align with both
association data (for a table) and genes data (for the plot).

Change a few layout examples so that instead of asking for a dummy
field, they ask for ":all". This only works on sources that bypass
parseData (like genes or burden tests) but is slightly less oaque
than just hard-coding "gene:gene" or "burden:dummy".

[ci skip]