Add new `element_selection` event
This is distinct from `element_clicked` in that it can distinguish
between selecting and deselecting. It provides a little more complexity
for when doing the extra filtering is worth it.

For example, one can build UIs that update an external table whenever
a gene is clicked. (and reset to an unfiltered view when the gene
is de-selected)