New widget to select LD population
- Add `dashboard_components` registry for complex and reusable buttons
- Add `set_state` widget that sets a value in plot.state
- Add `ldlz2_pop_selector` widget that changes LD population used
for whole plot

Button can optionally update its label text to show the selected value.

The options on the premade button are hardcoded for the 1000G
data in the UMich LD server. It is not part of the default layout,
because many deployments are still stuck on the old,
deprecated ap1v1 LD endpoint.