Remove the new memoization decorator, but keep in git history
Primary use case is scaling functions, but there's no clear cache key for the combination of parameters + field value

(eg, if we only memoize on field value, then two data layers with a different set of parameters would incorrectly see a cache hit for the same field value. Calculating a generic compound cache key is something of a performance hit, which defeats the point. The idea could be useful, but not in its current form.)