Allow data operations to mutate the data_layer instance that initiates them. (typically by changing layout, panel, panel.layout, etc)
In theory, could be used in future for things like "automatically create legend based on data". For now we document but don't promote this behavior heavily as the internals may evolve over time.

The previous way to achieve self-defining datalayer layouts (subclassing an existing layer) was often annoying, esp since CSS properties were scoped to the name of the parent class, which often caused the child class to render incorrectly. Data operations provide a safety valve to modify a well defined thing in a specific way.