Allow `set_state` to be used on panels, not just plots.
In theory, a clever developer could change a piece of state that only affected one adapter, and give the illusion of changing something panel specific. (even though plot.state is really global and shared) It could be neat and clever for advanced features.

In practice, someone might try to add the LD selector to every panel, and then get mad that the changes aren't isolated. We will want to watch user feedback closely to see how this expansion of features plays out.